refactor(learning): split employee dashboard into single-responsibility components
- Extract ContinueCard, RecommendedCard, ProgressRing, CategoryBreakdown,
  and AchievementsCard into dedicated files (one component per file)
- Slim dashboard orchestrator from 738 to ~210 lines
- Update barrel exports in components/index.ts
