+++ Frontend/apps/learning/src/components/index.ts
@@ -16,6 +16,11 @@ export { TrainingStatusTabs } from "./training-status-tabs";
 export { EnrolledTrainingCard } from "./enrolled-training-card";
 export { MyTrainingsList } from "./my-trainings-list";
 export { Dashboard } from "./dashboard/dashboard";
+export { ContinueCard } from "./dashboard/continue-card";
+export { RecommendedCard } from "./dashboard/recommended-card";
+export { ProgressRing } from "./dashboard/progress-ring";
+export { CategoryBreakdown } from "./dashboard/category-breakdown";
+export { AchievementsCard } from "./dashboard/achievements-card";
 export { AdminDashboard } from "./admin/admin-dashboard";
 export { EmployeeRow } from "./admin/employee-row";
 export { CompletionFunnel } from "./admin/completion-funnel";
